fix: Workarounds for soft deleted client in calendar
The client relation should be queried for events even if the client is soft-deleted.

Unfortunately, we loose the relation in the Livewire component, so it also is needed to re-query them unfortunately. When clicking the delete button, we also loose the client data, so 2 new property (name and address) is defined to keep the state.